move some code around and add takeover test
